# docs/contracts/sleep-care-phase1.openapi.yaml
openapi: 3.0.3
info:
  title: Sleep Care Phase 1 Contract
  version: P1-01-v1
  description: >-
    阶段一契约草案。只允许醒目标注的合成夹具；未批准医疗内容、真实通知、
    真实数据和面向康养用户的 AI 均保持禁用。成功与错误均使用 GVA
    {code,data,msg} 响应结构。
servers:
  - url: /api
    description: Web/Nginx 对外入口；后端与 Casbin 路径不含 /api 前缀
...
paths:
  /care/client-access/redeem:
    post:
      tags: [ClientAccess]
      operationId: redeemClientAccess
      summary: 一次性 grant 兑换康养用户受限会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemAccessRequest'
      responses:
        '200':
          $ref: '#/components/responses/RedeemAccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /care/client/tasks:
    get:
      tags: [ClientTasks]
      operationId: listClientTasks
      summary: 查询当前受限会话允许访问的任务
      description: 不接受 clientId、organizationId 或 assigneeId 扩大范围。
      security:
        - ClientSession: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/ClientTaskListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /care/client/tasks/{taskId}:
    get:
      tags: [ClientTasks]
      operationId: getClientTask
      summary: 查询受限会话内的任务详情
      security:
        - ClientSession: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          $ref: '#/components/responses/ClientTaskDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /care/client/tasks/{taskId}/questionnaire:
    get:
      tags: [ClientTasks]
      operationId: getClientTaskQuestionnaire
      summary: 获取任务冻结的问卷版本
      security:
        - ClientSession: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          $ref: '#/components/responses/ClientQuestionnaireResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /care/client/tasks/{taskId}/interactions:
    post:
      tags: [ClientTasks]
      operationId: recordClientTaskInteraction
      summary: 幂等记录打开、同意或开始交互
      description: 交互事件与通知回执、答卷提交和事项处理相互独立。
      security:
        - ClientSession: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordTaskInteractionRequest'
      responses:
        '200':
          $ref: '#/components/responses/ClientInteractionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /care/client/tasks/{taskId}/draft:
    put:
      tags: [ClientTasks]
      operationId: saveClientTaskDraft
      summary: 幂等保存任务草稿
      security:
        - ClientSession: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveDraftRequest'
      responses:
        '200':
          $ref: '#/components/responses/ClientDraftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /care/client/tasks/{taskId}/submit:
    post:
      tags: [ClientTasks]
      operationId: submitClientTask
      summary: 幂等提交任务答卷
      description: 提交成功不等于复核、处置或关闭完成。
      security:
        - ClientSession: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitTaskRequest'
      responses:
        '200':
          $ref: '#/components/responses/ClientSubmitTaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    EmployeeToken:
      type: apiKey
      in: header
      name: x-token
      description: GVA 后台员工令牌；不得用于康养用户访问。
    ClientSession:
      type: apiKey
      in: cookie
      name: gva_client_session
      description: 一次性 grant 兑换的 HttpOnly 受限会话。
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
        format: uint64
        minimum: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 调用方生成并在同一逻辑动作重试时复用的 UUID/ULID。
      schema:
        type: string
        minLength: 16
        maxLength: 128
  responses:
    BadRequest:
      description: 请求体无法解析或字段结构不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 员工令牌或康养用户受限会话无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 已认证但角色、组织、责任或敏感范围不允许
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: 未预期的基础设施异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RedeemAccessResponse:
      description: 兑换结果；成功时通过 Set-Cookie 写入受限会话
      headers:
        Set-Cookie:
          schema:
            type: string
          description: gva_client_session；生产必须 HttpOnly、Secure、SameSite=Lax。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RedeemAccessEnvelope'
...
    ClientTaskListResponse:
      description: 客户端受限任务分页；不包含客户或计划内部标识
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientTaskListEnvelope'
    ClientTaskDetailResponse:
      description: 客户端受限任务详情
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientTaskDetailEnvelope'
    ClientQuestionnaireResponse:
      description: 客户端冻结题目和可空草稿；不包含规则、审核或定义哈希
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientQuestionnaireEnvelope'
    ClientInteractionResponse:
      description: 客户端独立交互事实结果
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientInteractionEnvelope'
    ClientDraftResponse:
      description: 客户端草稿保存结果
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientDraftEnvelope'
    ClientSubmitTaskResponse:
      description: 客户端提交结果；不表示已处理
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientSubmitTaskEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [code, data, msg]
      properties:
        code:
          type: integer
          example: 41104
        data:
          nullable: true
        msg:
          type: string
          example: 无权执行该操作
    SuccessFields:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          enum: [0]
        msg:
          type: string
          example: success
    TaskExecutionStatus:
      type: string
      enum: [SCHEDULED, OPEN, IN_PROGRESS, SUBMITTED, CANCELLED]
    TaskTimingStatus:
      type: string
      enum: [NOT_OPEN, WITHIN_WINDOW, OVERDUE, EXPIRED]
    TaskReviewStatus:
      type: string
      enum: [NOT_READY, NOT_REQUIRED, PENDING, REVIEWING, REVIEWED, RETURNED]
...
    RedeemAccessRequest:
      type: object
      additionalProperties: false
      required: [grant]
      properties:
        grant:
          type: string
          minLength: 24
          maxLength: 512
          writeOnly: true
    RedeemAccessData:
      type: object
      required: [expiresAt, allowedTaskCount]
      properties:
        expiresAt:
          type: string
          format: date-time
        allowedTaskCount:
          type: integer
          minimum: 0
    RedeemAccessEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/RedeemAccessData'
    AnswerMap:
      type: object
      description: 以冻结问题编码为键；值类型由对应题型决定。
      additionalProperties:
        oneOf:
          - type: string
          - type: number
          - type: boolean
          - type: array
            items:
              type: string
    SaveDraftRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, answers]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
        answers:
          $ref: '#/components/schemas/AnswerMap'
    SubmitTaskRequest:
      type: object
      additionalProperties: false
      required: [expectedTaskVersion, source, answers]
      properties:
        expectedTaskVersion:
          type: integer
          minimum: 1
        source:
          type: string
          enum: [CLIENT_SELF]
        clientOccurredAt:
          type: string
          format: date-time
          description: 可选审计元数据；不能替代服务器提交时间。
        answers:
          $ref: '#/components/schemas/AnswerMap'
    RecordTaskInteractionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, interactionType]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        interactionType:
          type: string
          enum: [OPENED, CONSENTED, STARTED]
...
    ClientTaskSummary:
      type: object
      additionalProperties: false
      required: [id, dayCode, title, executionStatus, timingStatus, reviewStatus, openAt, dueAt, version, accessible, hasQuestionnaire]
      properties:
        id: {type: integer, format: uint64}
        dayCode: {type: string, pattern: '^D[1-9][0-9]*$'}
        title: {type: string}
        executionStatus: {$ref: '#/components/schemas/TaskExecutionStatus'}
        timingStatus: {$ref: '#/components/schemas/TaskTimingStatus'}
        reviewStatus: {$ref: '#/components/schemas/TaskReviewStatus'}
        openAt: {type: string, format: date-time}
        dueAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time, nullable: true}
        submittedAt: {type: string, format: date-time, nullable: true}
        version: {type: integer, minimum: 1}
        accessible: {type: boolean}
        hasQuestionnaire: {type: boolean}
    ClientTaskDetail:
      allOf:
        - $ref: '#/components/schemas/ClientTaskSummary'
        - type: object
          required: [opened, consented, started, canSaveDraft, canSubmit]
          properties:
            opened: {type: boolean}
            consented: {type: boolean}
            started: {type: boolean}
            canSaveDraft: {type: boolean}
            canSubmit: {type: boolean}
    ClientQuestionOption:
      type: object
      additionalProperties: false
      required: [code, label, order]
      properties:
        code: {type: string}
        label: {type: string}
        order: {type: integer, minimum: 1}
    ClientQuestion:
      type: object
      additionalProperties: false
      required: [code, type, title, required, order, validation, options]
      properties:
        code: {type: string}
        type:
          type: string
          enum: [SINGLE_CHOICE, MULTIPLE_CHOICE, TEXT, NUMBER, DATE, BOOLEAN]
        title: {type: string}
        required: {type: boolean}
        order: {type: integer, minimum: 1}
        validation:
          type: object
          additionalProperties: true
        options:
          type: array
          items: {$ref: '#/components/schemas/ClientQuestionOption'}
    ClientDraftState:
      type: object
      additionalProperties: false
      required: [version, answers, savedAt]
      properties:
        version: {type: integer, minimum: 1}
        answers: {$ref: '#/components/schemas/AnswerMap'}
        savedAt: {type: string, format: date-time}
    ClientQuestionnaire:
      type: object
      additionalProperties: false
      required: [id, title, purpose, expectedMinutes, taskVersion, questions, draft]
      properties:
        id: {type: integer, format: uint64}
        title: {type: string}
        purpose: {type: string}
        expectedMinutes: {type: integer, minimum: 0}
        taskVersion: {type: integer, minimum: 1}
        questions:
          type: array
          items: {$ref: '#/components/schemas/ClientQuestion'}
        draft:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ClientDraftState'
    ClientInteractionData:
      type: object
      additionalProperties: false
      required: [taskId, interactionType, executionStatus, taskVersion]
      properties:
        taskId: {type: integer, format: uint64}
        interactionType: {type: string, enum: [OPENED, CONSENTED, STARTED]}
        executionStatus: {$ref: '#/components/schemas/TaskExecutionStatus'}
        taskVersion: {type: integer, minimum: 1}
    ClientDraftData:
      type: object
      additionalProperties: false
      required: [taskId, version, savedAt]
      properties:
        taskId: {type: integer, format: uint64}
        version: {type: integer, minimum: 1}
        savedAt: {type: string, format: date-time}
    ClientSubmitTaskData:
      type: object
      additionalProperties: false
      required: [taskId, submissionId, executionStatus, reviewStatus, taskVersion, ruleHitIds, attentionCaseIds]
      properties:
        taskId: {type: integer, format: uint64}
        submissionId: {type: integer, format: uint64}
        executionStatus: {type: string, enum: [SUBMITTED]}
        reviewStatus: {$ref: '#/components/schemas/TaskReviewStatus'}
        taskVersion: {type: integer, minimum: 1}
        ruleHitIds:
          type: array
          items: {type: integer, format: uint64}
        attentionCaseIds:
          type: array
          items: {type: integer, format: uint64}
    ClientTaskPageData:
      type: object
      additionalProperties: false
      required: [page, pageSize, total, list]
      properties:
        page: {type: integer, minimum: 1}
        pageSize: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, format: int64, minimum: 0}
        list:
          type: array
          items: {$ref: '#/components/schemas/ClientTaskSummary'}
...
    ClientTaskListEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data: {$ref: '#/components/schemas/ClientTaskPageData'}
    ClientTaskDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data: {$ref: '#/components/schemas/ClientTaskDetail'}
    ClientQuestionnaireEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data: {$ref: '#/components/schemas/ClientQuestionnaire'}
    ClientInteractionEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data: {$ref: '#/components/schemas/ClientInteractionData'}
    ClientDraftEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data: {$ref: '#/components/schemas/ClientDraftData'}
    ClientSubmitTaskEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessFields'
        - type: object
          required: [data]
          properties:
            data: {$ref: '#/components/schemas/ClientSubmitTaskData'}
...